Fixed and cleaned up Autonomous double and triple cubes

+++ eclipse/Galaktika/src/org/usfirst/frc4905/Galaktika/groupcommands/AutoCombinedLeftRight.java
@@ -147,7 +147,7 @@ public class AutoCombinedLeftRight extends AutoCommand {
 	    }
 	}
 
-	protected void pickupFirstCubeFromScale(int deltaAngle) {
+	protected void pickupFirstCubeFromScale(double deltaAngle) {
 		turnDeltaAngle(deltaAngle);
 		moveElevatorToGroundHeight();
 		driveForward(53);
